requirements.txt
firebase-admin
streamlit
python-dotenv
instaloader
streamlit-option-menu
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
